#121: additional consolidation. depparser formatter submodule moved from cmds to depparser
