+++ src/ipbb/cmds/repo.py
@@ -15,7 +15,7 @@ from ..defaults import kSourceDir, kProjDir, kWorkAreaFile, kRepoSetupFile
 from ..depparser import Pathmaker
 from ..tools.common import mkdir
 from ..utils import DirSentry, findFileInParents, raiseError, formatDictTable
-from .formatters import DepFormatter
+from ..depparser import DepFormatter
 from .proj import info as proj_info
 from urllib.parse import urlparse
 from distutils.dir_util import mkpath
