+++ src/ipbb/depparser/__init__.py
@@ -2,3 +2,4 @@
 from ._fileparser import DepFileParser, DepFormatter, depfiletypes
 from ._pathmaker import Pathmaker
 from ._definitions import *
+from ._formatters import *
